Move a few more utility functions over into utils
